# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradeLedger API specification for Australian sole trader tradies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobSummary"
        "401":
          description: Unauthorized
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job with costs and invoices
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "401":
          description: Unauthorized
    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
  /jobs/{id}/status:
    put:
      operationId: updateJobStatus
      tags: [jobs]
      summary: Update job status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobStatusBody"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: jobId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceSummary"
        "401":
          description: Unauthorized
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
        "401":
          description: Unauthorized
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
        "401":
          description: Unauthorized
  /invoices/{id}/pay:
    post:
      operationId: markInvoicePaid
      tags: [invoices]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkInvoicePaidBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"
        "401":
          description: Unauthorized
  /invoices/{id}/remind:
    post:
      operationId: sendInvoiceReminder
      tags: [invoices]
      summary: Log reminder sent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reminder logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        name:
          type: string
        type:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        abn:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
      required:
        - id
        - name
...
    QuoteLineItem:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        quantity:
          type: string
          nullable: true
        unitCost:
          type: string
          nullable: true
        markupPercent:
          type: string
          nullable: true
        unitPrice:
          type: string
          nullable: true
        lineTotal:
          type: string
          nullable: true
        isInternal:
          type: boolean
          nullable: true
        sortOrder:
          type: integer
          nullable: true
      required:
        - id
        - description
...
    JobSummary:
      type: object
      properties:
        id:
          type: integer
        jobNumber:
          type: string
          nullable: true
        title:
          type: string
        status:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        total:
          type: string
          nullable: true
        scheduledDate:
          type: string
          nullable: true
        completedDate:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
        client:
          $ref: "#/components/schemas/Client"
          nullable: true
      required:
        - id
        - title
    Job:
      allOf:
        - $ref: "#/components/schemas/JobSummary"
        - type: object
          properties:
            description:
              type: string
              nullable: true
            subtotal:
              type: string
              nullable: true
            gstAmount:
              type: string
              nullable: true
            amountPaid:
              type: string
              nullable: true
            quoteId:
              type: integer
              nullable: true
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            expenses:
              type: array
              items:
                $ref: "#/components/schemas/Expense"
            invoices:
              type: array
              items:
                $ref: "#/components/schemas/InvoiceSummary"
            quoteLineItems:
              type: array
              items:
                $ref: "#/components/schemas/QuoteLineItem"
          required:
            - expenses
            - invoices
            - quoteLineItems
    CreateJobBody:
      type: object
      properties:
        clientId:
          type: integer
        quoteId:
          type: integer
        title:
          type: string
        description:
          type: string
        state:
          type: string
        status:
          type: string
        scheduledDate:
          type: string
        total:
          type: number
        subtotal:
          type: number
        gstAmount:
          type: number
      required:
        - title
    UpdateJobStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    InvoiceSummary:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        total:
          type: string
          nullable: true
        amountPaid:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        paidAt:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
        client:
          $ref: "#/components/schemas/Client"
          nullable: true
      required:
        - id
    InvoiceDetail:
      allOf:
        - $ref: "#/components/schemas/InvoiceSummary"
        - type: object
          properties:
            subtotal:
              type: string
              nullable: true
            gstAmount:
              type: string
              nullable: true
            paymentMethod:
              type: string
              nullable: true
            stripePaymentLink:
              type: string
              nullable: true
            sopaText:
              type: string
              nullable: true
            job:
              $ref: "#/components/schemas/JobSummary"
              nullable: true
            sentAt:
              type: string
              nullable: true
            reminder1SentAt:
              type: string
              nullable: true
            reminder2SentAt:
              type: string
              nullable: true
    CreateInvoiceBody:
      type: object
      properties:
        jobId:
          type: integer
        clientId:
          type: integer
        type:
          type: string
        total:
          type: number
        subtotal:
          type: number
        gstAmount:
          type: number
        dueDate:
          type: string
        stripePaymentLink:
          type: string
      required:
        - total
    MarkInvoicePaidBody:
      type: object
      properties:
        paymentMethod:
          type: string
        paidAt:
          type: string
      required:
        - paymentMethod
    Expense:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        jobId:
          type: integer
          nullable: true
        category:
          type: string
        vendor:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        amount:
          type: string
        gstClaimable:
          type: string
          nullable: true
        isGstClaimable:
          type: boolean
          nullable: true
        receiptUrl:
          type: string
          nullable: true
        expenseDate:
          type: string
        financialYear:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
      required:
        - id
        - category
        - amount
        - expenseDate
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
      required:
        - success